Add json-server for mockip up api

+++ src/app/employee/employee.component.ts
@@ -11,18 +11,29 @@ export class EmployeeComponent implements OnInit {
 
   employees:any = [];
 
+  keyword: any;
+
   constructor(public rest:RestService, private route: ActivatedRoute, private router: Router) { }
 
   ngOnInit() {
+    this.route.queryParams.subscribe(params => {
+      this.keyword = params['keyword'];
+    });
     this.getEmployees();
   }
 
   getEmployees() {
     this.employees = [];
-    this.rest.getEmployees().subscribe((data: {}) => {
+    console.log(this.keyword);
+    this.rest.getEmployees(this.keyword).subscribe((data: {}) => {
       console.log(data);
       this.employees = data;
     });
+  }
+
+  searchEmployees() {
+    this.router.navigate(['/employees'], { queryParams: { keyword: this.keyword } });
+    this.getEmployees();
   }
 
   add() {
@@ -30,13 +41,15 @@ export class EmployeeComponent implements OnInit {
   }
 
   delete(id) {
-    this.rest.deleteEmployee(id)
-      .subscribe(res => {
-          this.getEmployees();
-        }, (err) => {
-          console.log(err);
-        }
-      );
+    if (confirm('Are you sure?')) {
+      this.rest.deleteEmployee(id)
+        .subscribe(res => {
+            this.getEmployees();
+          }, (err) => {
+            console.log(err);
+          }
+        );
+    }
   }
 
 }
